requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
matplotlib~=3.6.0
seaborn==0.12.2

# Database
sqlite-utils==3.36

# Embeddings + Vector Search
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1

# NLP / Sentiment
textblob==0.18.0
vaderSentiment==3.3.2

# LLM Orchestration
langchain==0.2.12
langchain-community

openai